Almost there baby!
Added creation of two processes from the manager process and each player
process is started with a couple of run time arguments which make the
player processes either blue for player 1 and red for player 2 or vice
versa, randomly. The player process can be invoked directly from its
executable, without any command line arguments but in this case, the
processes starts but is not very interactive. Almost there baby, almost
there.